Issue #441 Support of 'GenBank' format - annotation part, gbk to gff - fixes
