+++ server/catgenome/src/main/java/com/epam/catgenome/component/MessageCode.java
@@ -52,7 +52,8 @@ public enum MessageCode {
     WRONG_BAM_INDEX_FILE("error.bam.index.file"),
     WRONG_NAME("name.already.exists"),
 
-    ERROR_GENBANK_FILE_READING("error.genbank.file.reading");
+    ERROR_GENBANK_FILE_READING("error.genbank.file.reading"),
+    ERROR_NO_QUALIFIERS("error.no.qualifiers");
 
     private final String code;
