+++ server/catgenome/src/main/java/com/epam/catgenome/manager/gene/GffManager.java
@@ -69,16 +69,14 @@ import com.epam.catgenome.manager.externaldb.bindings.ecsbpdbmap.PdbBlock;
 import com.epam.catgenome.manager.externaldb.bindings.ecsbpdbmap.Segment;
 import com.epam.catgenome.manager.externaldb.bindings.rcsbpbd.Record;
 import com.epam.catgenome.manager.externaldb.bindings.uniprot.Uniprot;
+import com.epam.catgenome.manager.genbank.GenbankManager;
 import com.epam.catgenome.manager.gene.parser.GeneFeature;
 import com.epam.catgenome.manager.gene.parser.GffCodec;
-import com.epam.catgenome.manager.gene.parser.StrandSerializable;
 import com.epam.catgenome.manager.gene.reader.AbstractGeneReader;
-import com.epam.catgenome.manager.gene.writer.Gff3FeatureImpl;
-import com.epam.catgenome.manager.gene.writer.Gff3Writer;
 import com.epam.catgenome.manager.parallel.ParallelTaskExecutionUtils;
 import com.epam.catgenome.manager.parallel.TaskExecutorService;
 import com.epam.catgenome.manager.reference.ReferenceGenomeManager;
-import com.epam.catgenome.manager.reference.io.GenbankUtils;
+import com.epam.catgenome.manager.genbank.GenbankUtils;
 import com.epam.catgenome.util.HistogramUtils;
 import com.epam.catgenome.util.IOHelper;
 import com.epam.catgenome.util.NggbIntervalTreeMap;
@@ -92,21 +90,12 @@ import htsjdk.samtools.util.IntervalTree;
 import htsjdk.tribble.AsciiFeatureCodec;
 import htsjdk.tribble.FeatureReader;
 import htsjdk.tribble.readers.LineIterator;
-import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.collections4.MapUtils;
 import org.apache.commons.io.FilenameUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.lang3.tuple.Pair;
-import org.biojava.nbio.core.sequence.DNASequence;
-import org.biojava.nbio.core.sequence.compound.NucleotideCompound;
-import org.biojava.nbio.core.sequence.features.DBReferenceInfo;
-import org.biojava.nbio.core.sequence.features.FeatureInterface;
-import org.biojava.nbio.core.sequence.features.Qualifier;
-import org.biojava.nbio.core.sequence.template.AbstractSequence;
-import org.biojava.nbio.genome.parsers.gff.Location;
-import org.jetbrains.annotations.NotNull;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.Assert;
@@ -119,9 +108,7 @@ import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Date;
-import java.util.HashMap;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.Map;
@@ -184,21 +171,14 @@ public class GffManager {
     @Autowired(required = false)
     private EhCacheBasedIndexCache indexCache;
 
+    @Autowired
+    private GenbankManager genbankManager;
+
     private static final String EXON_FEATURE_NAME = "exon";
 
     private static final String PROTEIN_CODING = "protein_coding";
 
     private static final int EXON_SEARCH_CHUNK_SIZE = 100001;
-
-    private static final String SOURCE = "GenBank";
-    private static final String PSEUDO = "pseudo";
-    private static final String DOT = ".";
-    private static final String LOCUS_TAG = "locus_tag";
-    private static final String OPERON = "operon";
-    private static final String CODON_START = "codon_start";
-    private static final String GENE = "gene";
-    private static final String NAME_ATTR = "Name";
-    private static final String PSEUDOGENE = "pseudogene";
 
     /**
      * Registers a gene file (GFF/GTF) in the system to make it available to browse. Creates Tabix index if absent
@@ -297,8 +277,9 @@ public class GffManager {
         if (GenbankUtils.isGenbank(path)) {
             String geneDir = fileManager.getGeneDir(geneFile.getId());
             Assert.notNull(geneDir, getMessage(MessageCode.RESOURCE_NOT_FOUND));
-            Path gffFilePath = Paths.get(geneDir, geneFile.getName() + GffCodec.GFF_EXTENSION);
-            genbankToGff(path, gffFilePath);
+            Path gffFilePath = Paths.get(geneDir,
+                    FilenameUtils.removeExtension(geneFile.getName()) + GffCodec.GFF_EXTENSION);
+            genbankManager.genbankToGff(path, gffFilePath);
             path = gffFilePath.toString();
             geneFile.setPath(path);
             request.setPath(path);
@@ -1221,75 +1202,6 @@ public class GffManager {
         return res;
     }
 
-    @SneakyThrows
-    public void genbankToGff(final String genebankFilePath, final Path gffFilePath) {
-        Map<String, DNASequence> dnaSequences = fileManager.readGenbankFile(genebankFilePath);
-        Assert.isTrue(!dnaSequences.isEmpty(), getMessage(MessageCode.ERROR_GENBANK_FILE_READING));
-        try (Gff3Writer gff3Writer = new Gff3Writer(gffFilePath)) {
-            for (Map.Entry<String, DNASequence> sequence : dnaSequences.entrySet()) {
-                String seqId = sequence.getValue().getAccession().getID();
-                Map<String, Integer> ids = new HashMap<>();
-                for (FeatureInterface<AbstractSequence<NucleotideCompound>, NucleotideCompound> f :
-                        sequence.getValue().getFeatures()) {
-                    Map<String, List<Qualifier>> qualifiers = f.getQualifiers();
-                    String type = GeneType.getType(f.getType(), qualifiers.containsKey(PSEUDO));
-                    Map<String, List<String>> attributes = new LinkedHashMap<>();
-                    List<String> value;
-                    if (type.equals(OPERON) && qualifiers.containsKey(OPERON)) {
-                        value = getQualifierValue(qualifiers, OPERON);
-                    } else if ((type.equals(GENE) || type.equals(PSEUDOGENE)) &&
-                            qualifiers.containsKey(LOCUS_TAG)) {
-                        value = getQualifierValue(qualifiers, LOCUS_TAG);
-                    } else {
-                        String idKey = (qualifiers.containsKey(LOCUS_TAG) ?
-                                qualifiers.get(LOCUS_TAG).get(0).getValue() :
-                                seqId) + DOT + type;
-                        int idValue = 0;
-                        if (ids.containsKey(idKey)) {
-                            idValue = ids.get(idKey) + 1;
-                            value = Collections.singletonList(idKey + idValue);
-                        } else {
-                            value = Collections.singletonList(idKey);
-                        }
-                        ids.put(idKey, idValue);
-                    }
-                    attributes.put("ID", value);
-                    if (f.getQualifiers().containsKey(GENE)) {
-                        attributes.put(NAME_ATTR, getQualifierValue(qualifiers, GENE));
-                    } else if (qualifiers.containsKey(LOCUS_TAG)) {
-                        attributes.put(NAME_ATTR, getQualifierValue(qualifiers, LOCUS_TAG));
-                    }
-                    attributes.putAll(qualifiersToAttr(qualifiers));
-
-                    Gff3FeatureImpl feature = new Gff3FeatureImpl(
-                            seqId.isEmpty() ? sequence.getValue().getOriginalHeader() : seqId,
-                            SOURCE,
-                            type,
-                            parseFeatureLocation(f.getSource()).getBegin(),
-                            parseFeatureLocation(f.getSource()).getEnd(),
-                            DOT,
-                            StrandSerializable.forValue(f.getLocations().getStrand().getStringRepresentation()),
-                            attributes.containsKey(CODON_START) ?
-                                    getQualifierValue(qualifiers, CODON_START).get(0) + 1 : DOT,
-                            attributes
-                    );
-                    try {
-                        gff3Writer.addFeature(feature);
-                    } catch (IOException e) {
-                        log.debug(e.getMessage(), e);
-                    }
-                }
-            }
-        }
-    }
-
-    private Location parseFeatureLocation(final String source) {
-        final String[] replacements = {"join(", ")", "complement(", "<", ">"};
-        final String[] replaceBy = {"", "", "", "", ""};
-        final String[] regions = StringUtils.replaceEach(source, replacements, replaceBy).split("\\..|,");
-        return new Location(Integer.parseInt(regions[0]), Integer.parseInt(regions[regions.length - 1]));
-    }
-
     private void loadExonsBackwards(int centerPosition, int viewPortSize, Chromosome chromosome,
                                     int intronLength, int featuresStart,
                                     final IntervalTree<Block> intervalTree,
@@ -1353,33 +1265,5 @@ public class GffManager {
             GeneFeature feature = iterator.next();
             totalLength = processExon(intervalTree, totalLength, feature, intronLength, centerPosition, true);
         }
-    }
-
-    private Map<String, List<String>> qualifiersToAttr(final Map<String, List<Qualifier>> qualifiers) {
-        final Map<String, List<String>> attributes = new LinkedHashMap<>();
-        for (Map.Entry<String, List<Qualifier>> qualifier : qualifiers.entrySet()) {
-            List<String> values = qualifier.getValue().stream().map(Qualifier::getValue).collect(Collectors.toList());
-            switch (qualifier.getKey()) {
-                case "db_xref":
-                    attributes.put("Dbxref", qualifier.getValue()
-                        .stream()
-                        .map(q -> ((DBReferenceInfo) q).getDatabase() + ":" + ((DBReferenceInfo) q).getId())
-                        .collect(Collectors.toList()));
-                    break;
-                case "note":
-                    attributes.put("Note", values);
-                    break;
-                default:
-                    attributes.put(qualifier.getKey(), values);
-                    break;
-            }
-        }
-        return attributes;
-    }
-
-    @NotNull
-    private List<String> getQualifierValue(final Map<String, List<Qualifier>> qualifiers, final String key) {
-        Assert.isTrue(qualifiers.containsKey(key), getMessage("No qualifiers with such a key"));
-        return qualifiers.get(key).stream().map(Qualifier::getValue).collect(Collectors.toList());
     }
 }
