+++ server/catgenome/src/main/resources/catgenome-messages.properties
@@ -266,3 +266,7 @@ error.blast.task.cant.be.deleted=Blast task can not be deleted
 error.blast.organisms=Blast task request should contain organisms OR excluded organisms
 error.blast.organisms.mapping=Can not map organism ids list to Taxonomy objects
 error.blast.database.not.found=Blast database not found
+
+#Genbank
+error.genbank.file.reading=Failed to read genbank file
+error.no.qualifiers=No qualifiers with such a key
