+++ server/ngb-cli/src/main/java/com/epam/ngb/cli/entity/BiologicalDataItemFormat.java
@@ -59,7 +59,10 @@ public enum BiologicalDataItemFormat {
     SEG_INDEX(12),
     MAF(13, false, true),
     MAF_INDEX(14),
-    VG(15);
+    VG(15),
+    GENBANK(16),
+    GBK(17),
+    GB(18);
 
     private long id;
     /**
@@ -105,6 +108,9 @@ public enum BiologicalDataItemFormat {
         EXTENSIONS_MAP.put("bdg", WIG);
         EXTENSIONS_MAP.put("bg", WIG);
         EXTENSIONS_MAP.put("bedGraph", WIG);
+        EXTENSIONS_MAP.put("genbank", GENBANK);
+        EXTENSIONS_MAP.put("gb", GB);
+        EXTENSIONS_MAP.put("gbk", GBK);
     }
 
     /**
